old-tests/submission/W3C/domhtml/HTMLTitleElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - TITLE</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-title-element.html#dom-title-text" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<br>
</p>



<script>
var domtest = async_test("The text attribute is the specified title as a string.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtext;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("title");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vtext = testNode.text;

      assert_equals(vtext,"NIST DOM HTML Test - TITLE");
       
    domtest.done();
}));

</script>
<p>The text attribute is the specified title as a string.      

    Retrieve the text attribute and examine its value.</p>
<div id="log"></div>
</body></html>
